+++ angular/dialogs/add_users/add_users.js
@@ -1,0 +1,17 @@
+(function(){
+    "use strict";
+
+    angular.module('app.controllers').controller('AddUsersCtrl', function($scope, DialogService){
+
+        $scope.save = function(){
+	        //do something useful
+            DialogService.hide();
+        };
+
+        $scope.hide = function(){
+        	DialogService.hide();
+        };
+
+    });
+
+})();