+++ app/Console/Kernel.php
@@ -15,6 +15,7 @@ class Kernel extends ConsoleKernel
     protected $commands = [
         'App\Console\Commands\Inspire',
         'App\Console\Commands\AngularFeature',
+        'App\Console\Commands\AngularDialog',
     ];
 
     /**